* Rewrite carousel to use hooks * Remove old unit tests * introduce jest * introduce yarn workspaces * remove .npmrc * update node version in .nvmrc * update README.md
