+++ .dockerignore
@@ -2,6 +2,7 @@
 .gitignore
 .editorconfig
 node_modules
+react-carousel/node_modules
 docs-www/node_modules
 *.DS_Store
 *.md
