+++ .nvmrc
@@ -1,1 +1,1 @@
-9
+12
