+++ README.md
@@ -391,7 +391,8 @@ render() {
 
 ## Unit tests
 ```
-yarn test:unit
+cd react-carousel
+yarn test
 ```
 
 ## E2E tests
@@ -402,9 +403,8 @@ yarn test:e2e
 
 ## Local running
 In order to run the docs/ demo locally:
-- `cd docs-www`
-- if you want to connect demo with the carousel source code, replace `__RC_ENV__` into `development` in https://github.com/brainhubeu/react-carousel/blob/master/docs-www/src/globalReferences.js#L2 and remove the `.babelrc` file in the root directory; otherwise, it will use the carousel code installed in `docs-www/node_modules`
-- `yarn develop`
+- if you want to connect demo with the carousel source code, replace `__RC_ENV__` into `development` in https://github.com/brainhubeu/react-carousel/blob/master/docs-www/src/globalReferences.js#L5
+- `yarn start-test-app`
 
 
 ## License
