+++ cypress.json
@@ -1,8 +1,3 @@
 {
-  "baseUrl": "http://localhost:8000",
-  "integrationFolder" : "test/cypress/integration",
-  "pluginsFile": "test/cypress/plugins/index.js",
-  "screenshotsFolder": "test/cypress/screenshots",
-  "supportFile": "test/cypress/support/index.js",
-  "videosFolder": "test/cypress/videos"
+  "baseUrl": "http://localhost:8000"
 }
