+++ cypress/integration/SwipeSlides.test.js
