+++ cypress/integration/Thumbnails.test.js
