+++ cypress/plugins/index.js
