+++ docs-www/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@brainhubeu/gatsby-docs-kit-start",
+  "name": "react-carousel-docs",
   "description": "Gatsby-Docs-Kit starter repository",
   "version": "0.0.37",
   "author": "Brainhub",
