+++ package.json
@@ -1,102 +1,16 @@
 {
-  "name": "@brainhubeu/react-carousel",
-  "version": "1.13.10",
-  "description": "Carousel component for React",
-  "engines": {
-    "npm": ">=6.14.3"
+  "private": true,
+  "workspaces": ["docs-www", "react-carousel"],
+  "devDependencies": {
+    "cypress": "^4.2.0",
+    "start-server-and-test": "^1.10.11"
   },
-  "files": [
-    "lib/*"
-  ],
   "scripts": {
-    "lint": "esw webpack.config.* src tools test --color",
-    "clean-dist": "rimraf ./lib && mkdir lib",
-    "prebuild": "npm run clean-dist",
-    "build": "node tools/build.js",
-    "cypress:open": "cypress open",
-    "cypress:run": "cypress run",
-    "test:unit": "mocha tools/testSetup.js \"./test/unit/*.js\" --exit",
+    "preinstall": "node -e \"if (process.env.npm_execpath.indexOf('yarn') === -1) throw new Error('Use yarn for installing: https://yarnpkg.com/en/docs/install')\"",
+    "start-test-app": "yarn workspace react-carousel-docs develop",
     "test:e2e": "start-server-and-test start-test-app http://localhost:8000 cypress:run",
     "test-dockerized:e2e": "docker-compose -f ./docker-compose.e2e.yml up --build --exit-code-from e2e-test",
-    "test:coverage": "NODE_PATH=example babel-node ./node_modules/.bin/isparta cover _mocha -- --require ./tools/testSetup.js \"./test/unit/*.js\" && open coverage/lcov-report/index.html",
-    "test:coverage:ci": "NODE_PATH=example babel-node ./node_modules/.bin/isparta cover _mocha --report lcovonly -- --require ./tools/testSetup.js \"./test/unit/*.js\"",
-    "start-test-app": "yarn --cwd docs-www develop",
-    "precommit": "npm run lint"
-  },
-  "author": "Brainhub",
-  "license": "MIT",
-  "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.8.7",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/polyfill": "^7.8.7",
-    "@babel/preset-env": "^7.8.7",
-    "@babel/preset-react": "^7.8.3",
-    "@babel/register": "^7.8.6",
-    "autoprefixer": "^9.7.4",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-transform-react-constant-elements": "^6.23.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "chai": "^4.2.0",
-    "chai-as-promised": "^7.1.1",
-    "chai-enzyme": "^1.0.0-beta.1",
-    "chalk": "^3.0.0",
-    "classnames": "^2.2.6",
-    "css-loader": "^3.4.2",
-    "cypress": "^4.2.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^6.8.0",
-    "eslint-config-brainhub": "^1.12.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-watch": "^6.0.1",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^5.1.0",
-    "gemini": "^7.5.2",
-    "html-reporter": "^5.13.0",
-    "html-webpack-plugin": "^3.2.0",
-    "husky": "^4.2.3",
-    "isparta": "^4.1.1",
-    "jsdom": "^16.2.1",
-    "lodash": "^4.17.15",
-    "mocha": "^7.1.1",
-    "node-sass": "^4.13.1",
-    "postcss-loader": "^3.0.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-test-renderer": "^16.13.1",
-    "rimraf": "3.0.2",
-    "sass-loader": "^8.0.2",
-    "sinon": "^9.0.1",
-    "sinon-chai": "^3.5.0",
-    "start-server-and-test": "^1.10.11",
-    "style-loader": "^1.1.3",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.42.0"
-  },
-  "peerDependencies": {
-    "react": ">0.14.0 || >15.0.0",
-    "react-dom": ">0.14.0 || >15.0.0"
-  },
-  "keywords": [
-    "react",
-    "carousel",
-    "react-carousel",
-    "react-component",
-    "component",
-    "image",
-    "image-gallery",
-    "image-slider",
-    "responsive",
-    "gallery",
-    "rwd"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/brainhubeu/react-carousel"
-  },
-  "main": "./lib/react-carousel.js"
+    "cypress:open": "cypress open",
+    "cypress:run": "cypress run"
+  }
 }
