+++ react-carousel/.eslintignore
