+++ react-carousel/.eslintrc
