+++ react-carousel/.istanbul.yml
