+++ react-carousel/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/test/**.test.js'],
+  moduleNameMapper: { '\\.(css|scss)$': '<rootDir>/test/__mocks__/styleMock.js' },
+  setupFilesAfterEnv: [
+    '<rootDir>/jest.setup.js',
+  ],
+};
