+++ react-carousel/jest.setup.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
+configure({ adapter: new Adapter() });
