+++ react-carousel/src/components/CarouselDots.js
