+++ react-carousel/src/components/CarouselItem.js
