+++ react-carousel/src/components/CarouselWrapper.js
