+++ react-carousel/src/constants/config.js
