+++ react-carousel/src/hooks/useEventListener.js
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from '../../docs-www/node_modules/react/index';
+import { useEffect, useRef } from 'react';
 
 function useEventListener(eventName, handler, element = window){
   // Create a ref that stores handler
