+++ react-carousel/src/hooks/usePrevious.js
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from '../../docs-www/node_modules/react/index';
+import { useEffect, useRef } from 'react';
 
 const usePrevious = value => {
   const ref = useRef();
