+++ react-carousel/src/index.js
