+++ react-carousel/src/styles/Arrows.scss
