+++ react-carousel/src/styles/Carousel.scss
