+++ react-carousel/src/styles/CarouselDots.scss
