+++ react-carousel/src/styles/CarouselItem.scss
