+++ react-carousel/tools/build.js
