+++ react-carousel/tools/chalkConfig.js
