+++ react-carousel/tools/startMessage.js
