+++ react-carousel/webpack.config.prod.js
