--- test/.eslintrc
@@ -1,8 +1,0 @@
-{
-  env: {
-    es6: true,
-    browser: true,
-    node: true,
-    mocha: true,
-  },
-}
